renaming notetype files to note_type
