+++ src/note_type.h
